refactored namings for stripe
