+++ src/screens/payment-methods/stripe/StripePaymentMethodCreationFormStyles.tsx
@@ -3,7 +3,7 @@ import { StyleSheet } from 'react-native';
 import ResponsiveStylesSheet from 'react-native-responsive-stylesheet';
 import { ScaledSheet } from 'react-native-size-matters';
 
-import Utils from '../../utils/Utils';
+import Utils from '../../../utils/Utils';
 
 /**
  *
